Parte Campaña
Sin revisar, ademas de refinar pequeños detalles en otras clases

+++ src/main/java/Extraccion.java
@@ -3,10 +3,10 @@ import java.time.LocalDate;
 public class Extraccion {
     private Donante voluntario;
     private LocalDate fechaExtraccion;
-    private float volumenExtraido;
+    private int volumenExtraido;
     private boolean seSintioMal;
    
-    public Extraccion (Donante voluntario, LocalDate fechaExtraccion, float volumenExtraido, boolean seSintioMal) {
+    public Extraccion (Donante voluntario, LocalDate fechaExtraccion, int volumenExtraido, boolean seSintioMal) {
         this.voluntario = voluntario;
         this.fechaExtraccion = fechaExtraccion;
         this.volumenExtraido = volumenExtraido;
@@ -16,7 +16,7 @@ public class Extraccion {
     //Getters
     public Donante getVoluntario() { return voluntario; }
     public LocalDate getFechaExtraccion() { return fechaExtraccion; }
-    public float getVolumenExtraido() { return volumenExtraido; }
+    public int getVolumenExtraido() { return volumenExtraido; }
     public boolean getSeSintioMal() { return seSintioMal; }
     
     //Setters
